Move util stuff into own dir
