+++ src/level_design/scene_editor.rs
@@ -1,10 +1,10 @@
-use crate::actions::{Actions, ActionsFrozen};
 use crate::file_system_interaction::game_serialization::{GameLoadRequest, GameSaveRequest};
 use crate::file_system_interaction::level_serialization::{WorldLoadRequest, WorldSaveRequest};
 use crate::level_design::spawning::{
     DelayedSpawnEvent, DuplicationEvent, GameObject, ParentChangeEvent,
     SpawnEvent as SpawnRequestEvent,
 };
+use crate::movement_gameplay::actions::{Actions, ActionsFrozen};
 use crate::GameState;
 use bevy::prelude::*;
 use bevy_egui::egui::ScrollArea;
