+++ src/movement_gameplay/actions.rs
@@ -1,7 +1,7 @@
 use bevy::input::mouse::MouseMotion;
 use bevy::prelude::*;
 
-use crate::actions::game_control::{get_movement, GameControl};
+use crate::movement_gameplay::actions::game_control::{get_movement, GameControl};
 use crate::GameState;
 
 mod game_control;
