+++ src/movement_gameplay/audio.rs
@@ -1,5 +1,5 @@
-use crate::actions::{set_actions, Actions};
 use crate::file_system_interaction::asset_loading::AudioAssets;
+use crate::movement_gameplay::actions::{set_actions, Actions};
 use crate::GameState;
 use bevy::prelude::*;
 use bevy_kira_audio::prelude::{Audio, *};
