+++ src/util/math.rs
